fix: resolvers linking
