--- apps/web/src/graphql/resolvers/query/index.ts
@@ -1,8 +1,0 @@
-import { IResolvers } from '@graphql-tools/utils'
-import { typeQueryResolvers } from '../type/queries'
-import { elementResolvers } from './element.resolvers'
-
-export const queryResolvers: IResolvers = {
-  ...elementResolvers,
-  ...typeQueryResolvers,
-}
