+++ apps/web/src/graphql/resolvers/queryResolvers.ts
@@ -1,8 +1,10 @@
 import { IResolvers } from '@graphql-tools/utils'
+import { elementQueryResolvers } from './element/query'
 import { typeQueryResolvers } from './type/queries'
 
 export const queryResolvers: IResolvers = {
   ...typeQueryResolvers,
+  ...elementQueryResolvers,
 }
 
 export default queryResolvers
