+++ apps/web/src/graphql/resolvers/type/mutations.ts
@@ -8,5 +8,3 @@ export const typeMutationResolvers: IResolvers = {
   deleteFieldEdge,
   upsertFieldEdge,
 }
-
-export default typeMutationResolvers
