+++ libs/frontend/modules/element/src/store/element/elementEndpoints.v2.ts
@@ -6,7 +6,6 @@ import {
   invalidatesAll,
   providesAll,
 } from '@codelab/frontend/model/infra/redux'
-import { AtomFragmentDoc } from '@codelab/frontend/modules/atom'
 import { Maybe } from '@codelab/shared/abstract/codegen-v2'
 import { Recipe } from '@reduxjs/toolkit/dist/query/core/buildThunks'
 import { api as generatedApi } from '../../graphql/element.endpoints.v2.graphql.gen'
@@ -45,7 +44,6 @@ export const GetElementsGraphGql = gql`
   ${PropFragmentDoc}
   ${HookFragmentDoc}
   ${PropMapBindingFragmentDoc}
-  ${AtomFragmentDoc}
 `
 
 const elementInjectedApi = generatedApi.injectEndpoints({
